[/API] rebuild: /API directory
