+++ API/API.sln
@@ -3,11 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 25.0.1700.5
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "API", "API\API.csproj", "{190B0D90-9777-47C1-AF1E-0E014028C99F}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OptimalAccessibility.Tests", "OptimalAccessibility.Tests\OptimalAccessibility.Tests.csproj", "{3B30FA65-0E27-4D35-B413-E2DC132AFB10}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "API", "OptimalAcessibility.API\API.csproj", "{3EB12E9B-E9E7-42A1-8F09-923413DDAE7C}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OptimalAccessibility.API", "OptimalAccessibility.API\OptimalAccessibility.API.csproj", "{02370A02-7AB0-41AE-B6B2-5C87650D554A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,6 +25,10 @@ Global
 		{3EB12E9B-E9E7-42A1-8F09-923413DDAE7C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3EB12E9B-E9E7-42A1-8F09-923413DDAE7C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3EB12E9B-E9E7-42A1-8F09-923413DDAE7C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{02370A02-7AB0-41AE-B6B2-5C87650D554A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{02370A02-7AB0-41AE-B6B2-5C87650D554A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{02370A02-7AB0-41AE-B6B2-5C87650D554A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{02370A02-7AB0-41AE-B6B2-5C87650D554A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
