+++ API/OptimalAccessibility.API/WeatherForecast.cs
@@ -1,0 +1,12 @@
+namespace OptimalAccessibility.API;
+
+public class WeatherForecast
+{
+    public DateTime Date { get; set; }
+
+    public int TemperatureC { get; set; }
+
+    public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
+
+    public string? Summary { get; set; }
+}
