--- API/OptimalAcessibility.API/appsettings.json
@@ -1,10 +1,0 @@
-﻿{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "AllowedHosts": "*"
-}
-
